requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0
colorlog>=6.7.0
